+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/DeveloperManager.java
@@ -12,7 +12,6 @@ import gov.healthit.chpl.dto.DeveloperACBMapDTO;
 
 public interface DeveloperManager {
 	public List<DeveloperDTO> getAll();
-	public List<DeveloperACBMapDTO> getAllTransparencies();
 	public DeveloperDTO getById(Long id) throws EntityRetrievalException;
 	public DeveloperDTO update(DeveloperDTO developer) throws EntityRetrievalException, JsonProcessingException, EntityCreationException;
 	public DeveloperDTO create(DeveloperDTO dto) throws EntityRetrievalException, EntityCreationException, JsonProcessingException;