+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/impl/DeveloperManagerImpl.java
@@ -3,7 +3,9 @@ package gov.healthit.chpl.manager.impl;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 import javax.mail.MessagingException;
@@ -61,41 +63,22 @@ public class DeveloperManagerImpl implements DeveloperManager {
 	@Override
 	@Transactional(readOnly = true)
 	public List<DeveloperDTO> getAll() {
-		Date begin = new Date();
-		logger.debug("Begin /developers call");
 		List<DeveloperDTO> allDevelopers = developerDao.findAll();
-		List<CertificationBodyDTO> availableAcbs = acbManager.getAllForUser(false);
-		if(availableAcbs == null || availableAcbs.size() == 0) {
-			availableAcbs = acbManager.getAll(true);
-		}
-		//someone will see either the transparencies that apply to the ACBs to which they have access
-		//or they will see the transparencies for all ACBs if they are an admin or not logged in
-		for(CertificationBodyDTO acb : availableAcbs) {
-			for(DeveloperDTO developer : allDevelopers) {
-				DeveloperACBMapDTO map = developerDao.getTransparencyMapping(developer.getId(), acb.getId());
-				if(map == null) {
-					DeveloperACBMapDTO mapToAdd = new DeveloperACBMapDTO();
-					mapToAdd.setAcbId(acb.getId());
-					mapToAdd.setAcbName(acb.getName());
-					mapToAdd.setDeveloperId(developer.getId());
-					mapToAdd.setTransparencyAttestation(null);
-					developer.getTransparencyAttestationMappings().add(mapToAdd);
-				} else {
-					developer.getTransparencyAttestationMappings().add(map);
-				}
-			}
-		}
-		Date end = new Date();
-		long millis = end.getTime() - begin.getTime();
-		logger.debug("End /developers in " + (millis/1000) + " seconds.");
-		return allDevelopers;
-	}
-
-    @Override
-    @Transactional(readOnly = true)
-	public List<DeveloperACBMapDTO> getAllTransparencies() {
-		List<DeveloperACBMapDTO> allDevelopers = developerDao.getAllTransparencyMappings();
-		return allDevelopers;
+		List<DeveloperACBMapDTO> transparencyMaps = developerDao.getAllTransparencyMappings();
+        Map<Long,DeveloperDTO> mappedDevelopers = new HashMap<Long,DeveloperDTO>();
+        for(DeveloperDTO dev : allDevelopers) {
+            mappedDevelopers.put(dev.getId(), dev);
+        }
+        for(DeveloperACBMapDTO map : transparencyMaps) {
+            if (map.getAcbId() != null) {
+                mappedDevelopers.get(map.getDeveloperId()).getTransparencyAttestationMappings().add(map);
+            }
+        }
+        List<DeveloperDTO> ret = new ArrayList<DeveloperDTO>();
+        for (DeveloperDTO dev : mappedDevelopers.values()) {
+            ret.add(dev);
+        }
+		return ret;
 	}
 
 	@Override