+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@ _Date TBD_
 * Added error messages for c1/2/3/4-CQM mismatch
 * Re-factored Certification Id web services to improve performance
 * Added statuses object that shows aggregate number of certified products associated with each developer and product. These objects will allow the website search page to filter on a developer/product's number of certified products that are active/retired/withdrawn/suspended/terminated.
+* Updated /developers call to be faster
 
 ### Bugs fixed
 * Pending 2015 products can now have CQM versions modified
@@ -38,7 +39,6 @@ _30 August 2016_
 * Task success avg was getting mixed up with Task errors avg
 * Task time deviation optimal avg was getting mixed up with task path deviation optimal avg
 * Test Task Participant Product & Professional experience was getting mixed up
->>>>>>> upstream/development
 
 ---